:Add images/tp build
This adapts the same methodology I used for my SAPSF LP2017 talk to keep
third-party resources out of the repository.  This is not only good from a
licensing perspective, but also good for the repo and programatically
defines how I derive the image displayed on my site from a source image (and
proves that it does not constitute a derivative work, as it is not
transformative).